Move thirdparty from pkg/api into internal
This package is only used for logic within pgrator, so no need to export
it
