+++ internal/controller/resourcecreator/opensearch.go
@@ -6,8 +6,8 @@ import (
 	"strconv"
 
 	"github.com/nais/pgrator/internal/config"
+	aiven_v1alpha1 "github.com/nais/pgrator/internal/thirdparty/aiven/v1alpha1"
 	"github.com/nais/pgrator/pkg/annotation"
-	aiven_v1alpha1 "github.com/nais/pgrator/pkg/api/thirdparty/aiven/v1alpha1"
 	v1 "github.com/nais/pgrator/pkg/api/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
