+++ internal/controller/resourcecreator/valkey.go
@@ -5,8 +5,8 @@ import (
 	"maps"
 
 	"github.com/nais/pgrator/internal/config"
+	aiven_v1alpha1 "github.com/nais/pgrator/internal/thirdparty/aiven/v1alpha1"
 	"github.com/nais/pgrator/pkg/annotation"
-	aiven_v1alpha1 "github.com/nais/pgrator/pkg/api/thirdparty/aiven/v1alpha1"
 	v1 "github.com/nais/pgrator/pkg/api/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
